Generating types and validator via optimized model (#17)
To generate proper types we want those types to be optimized before
generation. This PR introduces the usage of the type arena on which we
can apply transforms that will eventually turn the types into something
that we can easily turn into code.

Validation, mocks and everything is generated from the optimized types.

This is a big step forward but we are not there yet! Especially work
needs to be done on validation. We might end up with another type model
that is optimized for validation. Also we want to support edge cases
like an object with additional and properties. This is all to be done in
the future.