Use alternative types for validation (#38)
In order to efficiently create types we need to transform the original
schema quite a bit. But for validation, we don't need to mess with the
original schema too much. The original schema is actually better suited
for generating the validation logic, it is actually impossible to create
proper validation logic from the type model.

In this pr we create a different model that we use for validation.

- [x] add the ability to clone an arena
- [x] explode the not type
- [x] restructure specification creating logic
- [x] implement validation with the validation types
- [x] validation errors should have the correct path
- [x] debugging!
- [x] create mocks from validator type
- [x] implicit types
- [x] don't forget to handle implicit any, unknown and string types!
- [x] detect mockable types
- [x] only mock mockable types

- fixes https://github.com/LuvDaSun/JsonSchema42/issues/18
- fixes https://github.com/LuvDaSun/JsonSchema42/issues/39