splitting off core (#90)
A new project, in rust has all core stuff in it. It it exposed via ffi.
This project is then compiled top wasm so that it may be used in the
@jns42/core npm project. This project talks to the wasm binary via ffi.

(thank you js using keyword for making this easy to work with)

For every new language that wants to use the core, simply create a new
project and implemen the ffi interface. Implementing the interface may
be some work, but it's doable. The core rust project won't change as
more languages use it.

In node we now use the core as a wasm binary, but this could also be a
native dependency. This might be a better solutions for languages that
don't havce proper wasm support.

Wasm is preferred as we can ship one binary that will run everywhere.

- closes https://github.com/LuvDaSun/JsonSchema42/pull/89