Branded types, better validation (#129)
Branded types give a better experience than the current alias types. Two
branded types that have the same base type are not equal.

Imagine a type `UuidOrName`. This is a string that is either uuid
formatted or else it is any other string. With branded types we can do
something like

```
if(isUuid(value)) {
...
}
else {
...
}
```

without branded types, with type aliasses, then in the `else` arm
`value` would be `never`. This is because with an alias the `UuidOrName`
type would be `string`, just like the `Uuid` and `Name` types.

Also this pr includes better validation error messages and introduction
of a lib package.

---------

Co-authored-by: Elmer Bulthuis <elmerbulthuis@gmail.com>